+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Character actors gain -1 init bonus while not wearing armor.
+### Changed
+- Re-styled spell sheet.
 ### Fixed
 - Adjusted charisma modifiers were incorrectly applied to characters.